+++ src/pages/calendar/styles.js
@@ -16,13 +16,3 @@ export const DaysMonthStyles = styled.div`
     display: grid;
     grid-template-rows: repeat(7, 1fr);
 `;
-
-
-export const Cell = styled.div`
-    display: grid;
-    grid-template-columns: repeat(7, 1fr);
-`;
-
-export const DayStyles = styled.div`
-    background-color: ${(props) => props.currentDay === 'today' ? '#00838f' : '#fff'};
-`;
